verification external file add
