+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sportsportal",
-  "version": "0.5.11",
+  "version": "0.5.12",
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
